Propose config changes to replace env with more structured chain parameters (#19)
The src/config/index.js will act as a chain config selector for us to
handle more and more chains. Using env is good for the short-term but
not really scalable.

This PR is an example of how to use it in src/shibuya.js. The demo is
working with my changes.